refactor(cbft): insert tx with non-deterministic data
