+++ crates/consensus/authority/src/comet_bft/mod.rs
@@ -1,5 +1,5 @@
 pub mod abci;
-pub mod transactions_envelope;
+pub mod non_deterministic_data;
 
 // TODO remove mod once it's determined separate crate is the way to go
 // pub mod rpc;
